Solve day 11 part 2 of year 2022
The solution was to use product of all test dividers and take
modulus operation of item. End result is same even when the
item worry level is moderated with the product.

Had to cheat here a bit and look existing solutions. While looking
other solutions learned a bit of things which could be done bit
more elegantly - lambda functions and data classes could be neat.